I report a bug of the msgctl02 testcase and a patch for it. When you set /proc/sys/kernel/msgmnb to a value larger than 65536, msgctl02 always results FAIL because of a bug itself. An example when /proc/sys/kernel/msgmnb == 4194304, # cd ./testcases/kernel/syscalls/ipc/msgctl # ./msgctl02 msgctl02    1  FAIL  :  qs_buf.msg_qbytes value is not expected msgctl02    0  INFO  :  expected - 65535, received - 4194303 # This is because the variable "new_bytes" storing "expected" value above is defined as "unsigned short" type and cannot hold a value larger than 65535. It should be "msglen_t" type as same as the member msg_qnum of struct msqid_ds, but msg_qnum seems to be declared as "unsigned long int" type on some architecture like ia64, I recommend you to use "unsigned long int" for new_bytes. Signed-off-by: Seiichi Ikarashi <s.ikarashi@jp.fujitsu.com>.

+++ testcases/kernel/syscalls/ipc/msgctl/msgctl02.c
@@ -69,7 +69,7 @@ int msg_q_1 = -1;                      /* to hold the message queue id */
 
 struct msqid_ds qs_buf;
 
-unsigned short new_bytes;
+unsigned long int new_bytes;
 
 int main(int ac, char **av)
 {
